Eliminate products list search service (#186)
* Get rid of products-list-search-service
