+++ packages/headless-components/stores/src/react/core/ProductList.tsx
@@ -11,11 +11,6 @@ import {
   ProductServiceDefinition,
 } from '../../services/product-service.js';
 import {
-  ProductsListSearchService,
-  ProductsListSearchServiceConfig,
-  ProductsListSearchServiceDefinition,
-} from '../../services/products-list-search-service.js';
-import {
   CategoriesListService,
   CategoriesListServiceConfig,
   CategoriesListServiceDefinition,
@@ -29,8 +24,6 @@ export interface RootProps {
   children: React.ReactNode;
   /** Configuration for the ProductList service */
   productsListConfig: ProductsListServiceConfig;
-  /** Configuration for the ProductListSearch service */
-  productsListSearchConfig?: ProductsListSearchServiceConfig;
   /** Configuration for the CategoriesList service */
   categoriesListConfig?: CategoriesListServiceConfig;
 }
@@ -106,11 +99,6 @@ export function Root(props: RootProps): React.ReactNode {
           ProductsListServiceDefinition,
           ProductListService,
           props.productsListConfig,
-        )
-        .addService(
-          ProductsListSearchServiceDefinition,
-          ProductsListSearchService,
-          props.productsListSearchConfig,
         )}
     >
       {props.children}
